requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pytesseract
pdf2image
Pillow
pandas
xlsxwriter
openpyxl
google-genai
python-dotenv
pydantic